recommendation-service/requirements.txt
kafka-python
redis
pika
